check if file exists before deleting

+++ backend/download/signals.py
@@ -1,7 +1,6 @@
 import logging
 import os
 
-from django.conf import settings
 from django.db.models.signals import post_delete
 from django.dispatch import receiver
 from download.convert_csv import output_path
@@ -11,10 +10,12 @@ from .models import Download
 logger = logging.getLogger()
 
 def _try_remove_file(path):
-    try:
-        os.remove(path)
-    except:
-        logger.warning(f'Could not remove file: {path}', exc_info=True)
+    if os.path.exists(path):
+        try:
+            os.remove(path)
+        except:
+            logger.warning(f'Could not remove file: {path}', exc_info=True)
+
 
 @receiver(post_delete, sender=Download)
 def after_download_delete(sender, instance: Download, **kwargs):
@@ -26,5 +27,4 @@ def after_download_delete(sender, instance: Download, **kwargs):
             os.path.dirname(os.path.abspath(instance.filename)),
             instance.filename
         )
-        if os.path.exists(converted_path):
-            _try_remove_file(converted_path)
+        _try_remove_file(converted_path)
